Fix API documentation for 'breaches' query
For some reason, the 'breaches' API endpoint returns an empty array when no
breaches are found for the given query (when supplying a domain) rather than a
404 response. hibp, therefore, resolves the 'breaches' Promise with the empty
array, rather than undefined like the other endpoints. The JSDoc comment was
probably just copied from the others by mistake.

I'm tempted to normalize the 'breaches' behavior by resolving with undefined,
but it would require a major version bump as it would technically be a breaking
change to the hibp API.