Replace node-fetch with axios
Note: Sending a custom User-Agent header in the request has been (temporarily?)
dropped to facilitate browser support. Including our own User-Agent header
triggers a preflighted request (in the browser), which the current version of
the haveibeenpwned.com API (v2) does not support. I have reached out to the
author to see if this is intentional or just a CORS configuration mistake, and
may be able to re-enable this if the necessary changes are made to support
preflighted requests. For now, the User-Agent string will come from the user's
browser or from axios when running server-side. Obviously, there is no
functional difference whether we send our own User-Agent string or not - it's
just what the author requests in the API documentation.