Load tests with the Babel require hook and still get coverage
Referenced to get nyc/istanbul reporting properly with tests transpiled inline:
https://github.com/react-bootstrap/react-prop-types (as of v0.4.0)

Note on source maps:
Source map support comes with `mocha --compilers js:babel-register`, which
seems to enable source maps to work in tests but not in the compiled output.
Adding `import 'source-map-support/register'` to the src enables source maps
to work in compiled code but break in tests. Including source-map-support in
src and requiring source-map-support before babel-register when running mocha
(instead of passing the compilers option) appears to get both.

Referenced to get source maps working in test and compiled output:
https://github.com/evanw/node-source-map-support/issues/121

Notes on coverage:
nyc's "--all" option to include reports for files that are not required is
currently not working with babel-plugin-istanbul, but they're aware of the
matter. Also, including reporter settings in the nyc stanza of package.json
appear to trample command line options, so those were moved to the npm test
script instead.

Related issues:
https://github.com/istanbuljs/nyc/issues/333
https://github.com/istanbuljs/babel-plugin-istanbul/issues/4