Remove puppeteer to avoid downstream Chromium downloads
cf388136 downgraded the version of puppeteer because for some reason
I thought that was the culprit with regards to initiating a Chromium
download, but the older version does it too. Removing it entirely to
enhance the end-user/consumer experience at the cost of introducing
clunkiness to the test:umd script. It will now only run on CI unless
you manually install puppeteer and run it afterwards.