feat: add 'retryAfterSeconds' property to thrown rate limit errors
The haveibeenpwned.com API rate limits requests to prevent abuse. In the
event you get rate limited, the module will throw a custom
`RateLimitError` which will include a `retryAfterSeconds` property so
you know when you can try the call again (as a `number`, unless the
remote API did not provide one, in which case it will be `undefined` -
but that _should_ never happen).