test: refactor some tests
Some of the `fetchFromApi` tests were there out of a desire to be DRY,
but they've been duplicated and moved to their respective modules where
the behavior should actually be tested. As `fetchFromApi` is just an
implementation detail, the module-level behavior would need to remain
the same even if `fetchFromApi` was replaced.

The same argument could be made for some (if not all?) of the remaining
`fetchByApi` tests, but I'm not going to bother with those right now.